+++ src/encrypt/dto/result.ts
@@ -1,0 +1,19 @@
+
+import { ApiProperty } from "@nestjs/swagger";
+
+export class EncryptResult {
+    @ApiProperty({
+        description: 'Public key used by the gateway'
+    })
+    'public-key': string;
+    @ApiProperty({
+        description: 'Encrypted data'
+    })
+    hash: string;
+    @ApiProperty({
+        description: 'Encryption method (PSK-ECDSA or PSK-RSA)',
+        example: 'PSK-ECDSA'
+    })
+    method: string;
+}
+
